Create Class1_Activity1_transform-text-to-json-next-lang.json
1) Transform the following text into JSON:
“The "Next Lang" language academy is located at 123 main street in the city of Villa Carlos Paz. This academy offers a variety of courses, among which are:
"Basic English" Course: This introductory course is designed for beginners. It lasts for 4 weeks and is priced at $100.
"Advanced English" Course: This course is designed for advanced students. It lasts for 20 weeks and costs $400.
"Basic Italian" Course: This introductory course is designed for beginners. It lasts for 5 weeks and is priced at $120.
The academy also has a variety of contact telephone numbers, including the numbers 156734756 and 154456234. These telephone numbers are available to provide more information about the courses and resolve any queries that those interested may have.”

You can use the following online JSON editor: https://jsoneditoronline.org/classic/index.html